+++ eventful-rs/src/shard.rs
@@ -31,9 +31,7 @@ impl Shard {
     where
         F: FnOnce(&dyn Fn(T) -> ShardRc<T>) -> R + Send + 'static,
         R: Send + 'static,
-        T: Eventful<EventLoopHandleType = crate::ShardEventHandle>
-            + HasEvents<T::EventSetType>
-            + 'static,
+        T: Eventful + HasEvents<T::EventSetType> + 'static,
     {
         self.handle().bind_async(f)
     }
@@ -47,22 +45,11 @@ impl EventLoop for Shard {
     where
         F: FnOnce(&dyn Fn(T) -> ShardRc<T>) -> R + Send + 'static,
         R: Send + 'static,
-        T: Eventful<EventLoopHandleType = Self::HandleType> + HasEvents<T::EventSetType> + 'static,
+        T: Eventful + HasEvents<T::EventSetType> + 'static,
     {
         self.handle().bind(self.inner.owner(), f)
     }
     fn join(&self) -> Result<(), ShardError> {
         self.inner.join()
     }
-}
-#[macro_export]
-macro_rules! shard_std {
-    ($name:ident) => {
-        pub static $name: ::std::sync::LazyLock<$crate::shard::Shard> =
-            ::std::sync::LazyLock::new(|| $crate::shard::Shard::new(stringify!($name)));
-        type DefaultShardHandleType = $crate::shard::ShardEventHandle;
-        fn default_shard() -> &'static $crate::shard::Shard {
-            &$name
-        }
-    };
 }