+++ eventful-rs/src/tokio.rs
@@ -31,9 +31,7 @@ impl TokioShard {
     where
         F: FnOnce(&dyn Fn(T) -> ShardRc<T>) -> R + Send + 'static,
         R: Send + 'static,
-        T: Eventful<EventLoopHandleType = crate::ShardEventHandle>
-            + HasEvents<T::EventSetType>
-            + 'static,
+        T: Eventful + HasEvents<T::EventSetType> + 'static,
     {
         self.handle().bind_async(f)
     }
@@ -47,22 +45,11 @@ impl EventLoop for TokioShard {
     where
         F: FnOnce(&dyn Fn(T) -> ShardRc<T>) -> R + Send + 'static,
         R: Send + 'static,
-        T: Eventful<EventLoopHandleType = Self::HandleType> + HasEvents<T::EventSetType> + 'static,
+        T: Eventful + HasEvents<T::EventSetType> + 'static,
     {
         self.handle().bind(self.inner.owner(), f)
     }
     fn join(&self) -> Result<(), ShardError> {
         self.inner.join()
     }
-}
-#[macro_export]
-macro_rules! shard_tokio {
-    ($name:ident) => {
-        pub static $name: ::std::sync::LazyLock<$crate::tokio::TokioShard> =
-            ::std::sync::LazyLock::new(|| $crate::tokio::TokioShard::new(stringify!($name)));
-        type DefaultShardHandleType = $crate::tokio::TokioShardHandle;
-        fn default_shard() -> &'static $crate::tokio::TokioShard {
-            &$name
-        }
-    };
 }