+++ eventful-rs/tests/runtime.rs
@@ -25,7 +25,7 @@ impl TestState {
 }
 impl Eventful for TestState {
     type EventSetType = ();
-    type EventLoopHandleType = ShardEventHandle;
+    type Shard = eventful_rs::DynamicShard;
 }
 impl HasEvents<()> for TestState {
     fn events(&self) -> &Arc<()> {